requirements.txt
python-dotenv
pydub
openai
